Move sync-specific types, helpers and implementations into the sync module
